serve: fix a potential crash when adding a tuner/feeder after one has been removed

+++ libdvbtee_server/serve.h
@@ -30,8 +30,8 @@
 
 #define SERVE_DEFAULT_PORT 64080
 
-typedef std::map<uint8_t, tune*> tuner_map;
-typedef std::map<uint8_t, feed*> feeder_map;
+typedef std::map<unsigned int, tune*> tuner_map;
+typedef std::map<unsigned int, feed*> feeder_map;
 
 class serve;
